requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
Pillow>=9.5.0
python-dotenv>=1.0.0
psutil>=5.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
qrcode[pil]>=7.4.0
playwright>=1.40.0
cryptography>=41.0.0
groq>=0.4.0
# FIXED: Added aiosqlite for async database operations (Critical Issue #1)
aiosqlite>=0.19.0
# Архитектура: Postgres + Redis
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=2.0.0
prometheus_client>=0.18.0
brotli>=1.0.9  # For brotli decompression support